create single app config and move backend files
