+++ frontend/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  API_URL: 'http://localhost:3000/api/v1'
+  API_URL: '/api/v1'
 };
