+++ local-db/docker-compose.yaml
@@ -1,7 +1,7 @@
 version: '3.3'
 services:
   db:
-    image: mysql:5.7
+    image: mysql:8.0.23
     restart: always
     environment:
       MYSQL_DATABASE: 'shopmate'
