+++ src/routes/api/categories.ts
