+++ src/routes/api/products.ts
