+++ src/routes/api/search.ts
